Create initial clone of Parse Todo app
The Todo app is a useful starting point to build off of.  We need basic
user authentication, and limited input.  The first steps will be to
strip down the app to the basics, until it just serves as a checkbox
poll for days of the week.  Picking friends to get matched with will
come later.