Rewrite most of the front-end to be prettier and have nicer code
The design is updated to have a single table, with one mutable row for the current user, and immutable rows for the rest. It uses Material Design Lite (http://www.getmdl.io/started/index.html) to implement some reasonable-ish styling, although the login page remains unstyled for now.

The JS was simplified to no longer be a twisted version of TodoMVC. This was necessary anyway since the HTML structure of the page changed so much.

The CSS is now inline (generally better for small amounts of crucial CSS, to avoid blocking rendering on a HTTP request), and we use Underscore.js from a CDN. The images are no longer necessary.

A few TODOs are left sprinkled through the code, many having to do with changing the backend data model to be a bit easier to work with.